Update inspector panels design
Ref https://github.com/webstudio-is/webstudio-builder/issues/1644

- added instance info to all panels
- fixed style sources design
