+++ apps/builder/app/builder/features/settings-panel/settings-panel.tsx
@@ -1,11 +1,5 @@
 import { useStore } from "@nanostores/react";
-import {
-  Flex,
-  Label,
-  theme,
-  ScrollArea,
-  InputField,
-} from "@webstudio-is/design-system";
+import { Flex, Label, theme, InputField } from "@webstudio-is/design-system";
 import { getComponentMeta } from "@webstudio-is/react-sdk";
 import { selectedInstanceStore } from "~/shared/nano-states";
 import { useSettingsLogic } from "./use-settings-logic";
@@ -18,26 +12,24 @@ export const SettingsPanel = () => {
   }
   const label = getComponentMeta(selectedInstance.component)?.label;
   return (
-    <ScrollArea>
-      <Flex
-        gap="1"
-        direction="column"
-        grow
-        css={{ px: theme.spacing[9], py: theme.spacing[9] }}
-      >
-        <Label>Instance Name</Label>
-        <InputField
-          /* Key is required, otherwise when label is undefined, previous value stayed */
-          key={selectedInstance.id}
-          onBlur={handleBlur}
-          onKeyDown={handleKeyDown}
-          placeholder={label}
-          defaultValue={selectedInstance.label}
-          onChange={(event) => {
-            setLabel(event.target.value.trim());
-          }}
-        />
-      </Flex>
-    </ScrollArea>
+    <Flex
+      gap="1"
+      direction="column"
+      grow
+      css={{ px: theme.spacing[9], py: theme.spacing[9] }}
+    >
+      <Label>Instance Name</Label>
+      <InputField
+        /* Key is required, otherwise when label is undefined, previous value stayed */
+        key={selectedInstance.id}
+        onBlur={handleBlur}
+        onKeyDown={handleKeyDown}
+        placeholder={label}
+        defaultValue={selectedInstance.label}
+        onChange={(event) => {
+          setLabel(event.target.value.trim());
+        }}
+      />
+    </Flex>
   );
 };
