+++ packages/design-system/src/components/combobox.tsx
@@ -95,10 +95,14 @@ export const ComboboxListboxItem = forwardRef(ListboxItemBase);
 
 export const Combobox = Popper;
 
+export const StyledPopperContent = styled(PopperContent, {
+  zIndex: theme.zIndices.max,
+});
+
 export const ComboboxContent = forwardRef(
   (props: ComponentProps<typeof PopperContent>, ref: Ref<HTMLDivElement>) => (
     <Portal>
-      <PopperContent ref={ref} {...props} />
+      <StyledPopperContent ref={ref} {...props} />
     </Portal>
   )
 );
